refactor(deliverables): 回流成品业务逻辑到 deliverables 特性
