+++ pinvou3-app/src-tauri/src/features/mod.rs
@@ -1,6 +1,7 @@
 pub mod assistant;
 pub(crate) mod codex_acp;
 pub(crate) mod connectors;
+pub(crate) mod deliverables;
 pub(crate) mod dependencies;
 pub mod feedback;
 pub mod files;
